Add CallSoundPlayer
